Create Lemonade change
Ninja is a lemonade seller and all the lemonade costs 5 rupees for each product. The amount paid by the each customer will be 5,10 or 20 rupees.
You are given an array bill of length N, where bill[i] represents the amount paid by the ith customer. You must return the extra amount to the customers in the order they appear in the bill. 
Given the array bill , you must return whether or not Ninja can return the correct change to all the customers
eg: bill=[5,5,5,10,10,20]; op=True