Create Immediate Smaller Element
You are given an integer array 'a' of size 'n'.


For each element in the array, check whether the immediate right element of the array is smaller or not.


If the next element is smaller, update the current index to that element. If not, then -1. The last element does not have any elements on its right.