Create Sum of subarray min
You are given an array 'arr' of length ‘N’.

Let ‘X’ be the minimum element of any contiguous subarray of ‘arr’.

You need to return the sum of 'X' over all the contiguous subarrays of 'arr'. Since the answer may be large, you should return it modulo 10^9+7.

A contiguous subarray is a subset of consecutive elements from an array where the elements are adjacent and appear in the same order as in the original array.