Create Max size rect sub matrix with all 1's
You are given an 'N' * 'M' sized binary-valued matrix 'MAT, where 'N' is the number of rows and 'M' is the number of columns. You need to return the maximum size (area) of the sub matrix which consists of all 1’s i.e. the maximum area of a sub matrix in which each cell has only the value ‘1’.